backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart
aiofiles
requests
beautifulsoup4
transformers
torch
gtts
python-docx
pypdf
lxml
pdfminer.six
